added job endpoints (create , read , update)

+++ quickcollab/src/main/java/com/quickcollab/model/Conversation.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class Conversation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer conversationId;
+    private Long conversationId;
 
     @OneToMany
     private List<User> members;
